requirements.txt
ai21
streamlit
python-dotenv
google
beautifulsoup4